refactor to make easier to iterate on
