+++ packages/holodeck/server/bun-worker.js
@@ -1,0 +1,3 @@
+import { startWorker } from './bun.js';
+
+startWorker();
