+++ packages/holodeck/server/node-worker.js
@@ -1,0 +1,3 @@
+import { startWorker } from './node.js';
+
+startWorker();
