--- packages/holodeck/server/start-node.js
@@ -1,3 +1,0 @@
-import { startNodeServer } from './index.js';
-
-startNodeServer();
