--- packages/holodeck/server/worker.js
@@ -1,3 +1,0 @@
-import { startWorker } from './index.js';
-
-startWorker();
